@@ -12,10 +12,10 @@
     "eslint-plugin-jsx-a11y": "^2.2.2",
     "eslint-plugin-react": "6.3.0",
     "jest": "15.1.1",
-    "redux-devtools": "^3.3.1",
-    "redux-devtools-chart-monitor": "^1.5.1",
-    "redux-devtools-dock-monitor": "^1.1.1",
-    "redux-devtools-log-monitor": "^1.0.11",
+    "redux-devtools": "^3.4.0",
+    "redux-devtools-chart-monitor": "^1.6.1",
+    "redux-devtools-dock-monitor": "^1.1.2",
+    "redux-devtools-log-monitor": "^1.3.0",
     "webpack-bundle-size-analyzer": "^2.7.0"
   },
   "dependencies": {
@@ -67,18 +67,18 @@
     "pretty-error": "^2.0.2",
     "promise": "7.1.1",
     "prop-types": "^15.5.8",
-    "react": "^15.6.0",
+    "react": "^16.0.0",
     "react-addons-css-transition-group": "^15.4.2",
-    "react-bootstrap": "^0.30.8",
+    "react-bootstrap": "^0.31.3",
     "react-breadcrumbs-dynamic": "^1.0.6",
-    "react-dom": "^15.4.2",
-    "react-helmet": "^3.3.0",
+    "react-dom": "^16.0.0",
+    "react-helmet": "^5.2.0",
     "react-hot-loader": "^3.0.0-beta.6",
     "react-redux": "^5.0.2",
     "react-router": "^3.0.2",
     "react-router-redux": "^4.0.8",
     "react-router-scroll": "^0.4.1",
-    "react-toggle-button": "^2.1.0",
+    "react-toggle": "^4.0.2",
     "redbox-react": "^1.3.3",
     "redux": "^3.6.0",
     "redux-connect": "^5.1.0",
