@@ -6,10 +6,10 @@ let DevTools = null;
 
 if (__DEVELOPMENT__ && __DEVTOOLS__) {
   const { createDevTools } = require('redux-devtools');
-  const ChartMonitor = require('redux-devtools-chart-monitor').default;
+  //const ChartMonitor = require('redux-devtools-chart-monitor').default;
   const DockMonitor = require('redux-devtools-dock-monitor').default;
   const LogMonitor = require('redux-devtools-log-monitor').default;
-  const SliderMonitor = require('redux-slider-monitor');
+  //const SliderMonitor = require('redux-slider-monitor');
 
   DevTools = createDevTools(
     <DockMonitor
@@ -18,8 +18,8 @@ if (__DEVELOPMENT__ && __DEVTOOLS__) {
       changePositionKey="ctrl-p"
       changeMonitorKey="ctrl-m">
         <LogMonitor />
-        <SliderMonitor keyboardEnabled />
-        <ChartMonitor />
+        {/*<SliderMonitor keyboardEnabled />*/}
+        {/*<ChartMonitor />*/}
     </DockMonitor>
   );
 }
